yolo-clothes/requirements.txt
fastapi
uvicorn[standard]
pymongo[srv]
python-jose[cryptography]
passlib[bcrypt]
python-multipart
Pillow
torch
opencv-python-headless
numpy